Changed comments on dataset load and preprocess.
Changed the encoding process to be more dynamic, allowing for easier use of encoding multiclass datasets in the future.

add a separate processing function for the AC data since the model deals in multicategorical data with with a larger dimension

separated the original preprocess function with the current version.

Reorganize structures for discriminator and generator

took out redundant imports in nidsModelCentralTrainingConfig.py